backend/requirements.txt
# Atlas AI Command Center Backend - Python 3.11+
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
websockets==12.0
psycopg2-binary==2.9.9
google-generativeai==0.8.0
httpx==0.27.0
